placeholder test: follow redirects, check for 200
When I tried running the test again with Tomcat running, it failed.
Checking the response status, it was returning HTTP 302 Found:

$ curl -I http://localhost:8080/strcalc

  HTTP/1.1 302
  Location: /strcalc/
  Transfer-Encoding: chunked
  Date: Fri, 27 Oct 2023 19:35:07 GMT

At first I added the trailing slash, and it passed, but I figure it
might be worth making sure the client follows redirects instead.

(Of course, to execute the test, I had to comment out the @Disabled
annotation temporarily.)