Add /strcalc HTML and Hamcrest matchers
Now that the small, medium, and large test targets are working, we're
close to ready to add large Selenium WebDriver tests. First, we serve a
proper HTML page from /strcalc and move the existing servlet endpoint
to /strcalc/add. Then, we use a Hamcrest matcher to check for the
presence of "Hello, World!" in the initial page.

More details in the README.