Break `test-*` task dependency on `test` task
Using the `smallTests` object corresponding to the actual `test` task
created a dependency on that task by any task applying
`addCommonTestSuiteConfiguration()`. This change uses the literal `val
test by ...` expression from the following to break the dependency:

- https://docs.gradle.org/8.4/userguide/upgrading_version_8.html#test_task_default_classpath

I also didn't understand the `val test by ...` syntax earlier. I still
wouldn't say I _fully_ understand it, but the name `test` is referring
to a specific element in `testing.suites`. The expression originally
caused an error when I tried to call it `defaultTestSuite`.

The `JvmTestSuite` class is still marked as incubating for now, which is
part of why I avoided using it for test sizes:

- https://docs.gradle.org/current/userguide/java_testing.html#test_grouping
- https://docs.gradle.org/current/userguide/jvm_test_suite_plugin.html
- https://docs.gradle.org/current/javadoc/org/gradle/api/plugins/JvmTestSuitePlugin.html
- https://docs.gradle.org/current/javadoc/org/gradle/api/plugins/jvm/JvmTestSuite.html
- https://docs.gradle.org/current/javadoc/org/gradle/api/Incubating.html
- https://docs.gradle.org/current/userguide/feature_lifecycle.html#sec:incubating_state

But if Gradle is documenting that this is the preferred idiom to set
`Test` target parameters, so be it.