Emit paths to generated coverage reports
This is similar in spirit to emitting merged test result paths in commit
0435b0627512a893930158e4bcf801514893cad9.

The nice feature here is that the JacocoReport has executionData
dependencies on JaCoCo *.exec files. The tasks won't run and output
anything on the console if the report output file is up to date with the
input *.exec files.