Add @vitest/coverage-istanbul 1.0.0-beta.4
This enables running tests with Istanbul as an alternative to V8
coverage:

  pnpm test -- --coverage.enabled --coverage.provider=istanbul

It's also required to collect code coverage from tests running in a
browser, as @vitest/coverage-v8 will not work, per the message from
commit 98c3ce1bb285180a6f9d2578ab16ef8ec9d37529.

---

Also following up on that commit message: I have found a hack for
getting browser coverage for pages loaded via the browser's window.open.
It's not pretty or perfect, but it works well enough for my purposes.
That'll be the next commit.

However, I wanted to get the @vitest/coverage-istanbul commit in
separately. I think I'll always want to keep this package even if I
eventually revert the hack I'm about to commit.