Integrate Handlebars into frontend JavaScript
This required creating a new Rollup plugin to precompile the Handlebars
templates as part of the Vite build. I intend to break this plugin out
into a separate repository, which I may publish. I'm getting inspiration
for this plugin from:

- https://github.com/benjilegnard/rollup-plugin-handlebars
- https://github.com/mixmaxhq/rollup-plugin-handlebars-plus

This commit adds a new frontend/components directory, in which I've
added a new "Placeholder" component consisting of placeholder.js and
placeholder.hbs. It seems I'm heading in the direction of emulating
popular modern frameworks like React or Vue, but using Handlebars
instead. We'll see how this goes, as I'm about ready to add the
actual String Calculator details now.

As for why Handlebars, or even Tomcat, it's because I'm developing this
example for a specific team to whom I'm going to present it soon. These
are the technologies they're using. (I should add this to the README at
some point...)

The example project as a whole is becoming a bit of a Rube Goldberg
machine, but that's kind of the point. I want it to illustrate how to
structure and test a production system, but without requiring too much
time to learn this specific system. The actual business logic of the
String Calculator is secondary to showing how to design for testability
and how to implement a comprehensive testing strategy.

I like to think of it as trying to narrow the "imagination gap" between
what's possible and what folks are aware is possible.