Set TestTomcat baseDir explicitly
This is in preparation for possibly launching a TestTomcat per test
method, instead of per test fixture/suite.

I found out that trying to dynamically create a baseDir for each
instance led to the first instance's baseDir reappearing. I haven't been
able to see clearly how this is happening, but it seems that:

1. The baseDir isn't strictly necessary in our case, since we're not
   running JSPs.
2. Using the same build/test-tomcat baseDir for all instances and having
   each instance delete it when finished isn't breaking anything.