Add hasContentType matcher, HTTP status consts
This matcher removes a lot of boilerplate from checking the Content-Type
of a HttpResponse. It's based on the Hamcrest API with the help of these
tutorials:

- https://hamcrest.org/JavaHamcrest/tutorial
- https://www.baeldung.com/hamcrest-custom-matchers

This could be generalized to handle any header.  However, I want to keep
it straightforward for now, as a good basic example of how to write a
custom matcher.

Also took the opportunity to replace bare HTTP status codes (200, et.
al.) with constants defined on jakarta.servlet.http.HttpServletResponse.