Try refining Run Tests GitHub Action again
PR #7 triggered the Run Tests action again, but did so multiple times.

I want all pushes to `main` to trigger builds, as well as updates to
pull requests, but not pushes to other branches. And I want each action
to run only once per PR.

This is an attempt to get the right thing to happen based on advice
from:

- https://stackoverflow.com/a/65096459
- https://github.com/orgs/community/discussions/26276#discussioncomment-3251140