Refactor Handlebars precompiler as PluginImpl
This is in preparation for supporting partial templates. As it became
apparent that compiledModule() would need more information, it made
sense to create a PluginImpl class, versus passing a bunch of
arguments.

Though it's a little more code, I think what's happening inside the
plugin is even more understandable now.