Parse Handlebars AST, export default Template
This is in preparation for adding support for Handlebars partials.

First, it now parses the Abstract Syntax Tree using Handlebars.parse(),
then passes the AST into Handlebars.precompile() (instead of the
original code). This will allow a Handlebars.Visitor to walk the AST and
discover the names of partials.

Second, the compiled module now exports a `Template` variable, instead
of exporting the result of Handlebars.template() directly. This will
enable compiledModule() to emit a call to Handlebars.registerPartial()
on the `Template` as necessary.

Also added the potential to pass Handlebars compiler options through via
options.compiler, and added the IMPORT_HELPERS constant.