Emit Handlebars source map by default
Reading the following again, emitting the source map should be the
default unless explicitly configured via `sourceMap: false` or
`sourcemap: false`:

- https://rollupjs.org/plugin-development/#source-code-transformations
- https://rollupjs.org/troubleshooting/#warning-sourcemap-is-likely-to-be-incorrect

So now, if the plugin options contain `sourceMap: false` or `sourcemap:
false`, `vite build --sourcemap` will emit the "Sourcemap is likely to
be incorrect" warning.

Note the property can be "sourcemap" or "sourceMap", because each of the
two above documents use a different version.