Set Handlebars #adjustSourceMap in PluginImpl()
This hoists the decision of whether or not to adjust source maps into
the PluginImpl constructor(). We usually always will or always won't
generate them, so this makes compile() read even more imperatively.

Also performed some other light refactoring in the PluginImpl
constructor.

---

It may be conceivable one day that someone would want to filter whether
or not to produce source maps based on a file pattern. It would be
pretty easy to apply such a filter in the PluginImpl.#compilerOpts
lambda PluginImpl.#compilerOpts, and to add a check to #adjustSourceMap,
without changing compile(). This is because compile() uses a
destructuring assignment on the result from Handlebars.precompile() to
assign the `tmpl` and `map` constants:

- https://developer.mozilla.org/docs/Web/JavaScript/Reference/Operators/Destructuring_assignment

When there's a `srcName` field in the options, Handlebars.precompile()
returns a `{code, map}` object. When there isn't, it returns a string.

- https://handlebarsjs.com/api-reference/compilation.html#handlebars-precompile-template-options

The destructuring assignment ensures that `tmpl` and `srcMap` are set
accordingly either way. It's like a conditional statement, but reads
more like an imperative one.

So making adding a filter to decide which files get source maps or not
would be easy. However, I'm not sure how common a use case that would
be, so I'll leave it out for now.