Make emitted Handlebars plugin code smaller
The helpers module now exports an (unnamed) default function that takes
a function created by Handlebars.template and wraps it in a new
function. The returned function both executes the input function and
renders its result as an HTMLCollection.

This doesn't change observable behavior, but makes the code emitted for
each Handlebars template module slightly smaller.