WIP: Get closer to eval of ESM in JSDOM context
It turns out I can pass through the functions used to resolve module IDs
and to fetch the module code after all. It even returns the Handlebars
code generated from rollup-plugin-handlebars-precompile. The problem is
that the code has already been transformed from an ES Module into a
format that Vitest will invoke as an IIFE.

This means that some standard ES Module keywords—including import and
export—have been replaced by Vitest-defined variables.
ViteNodeRunner.runModule() takes a context object that replaces these
variables when passed in as arguments to the IIFE.

So close now... If there's a way to stop that final transformation,
while keeping the rest, I could get the proof of concept working. (Or I
could get it working with a module that imports nothing, at least.)