WIP: Attempt to pass through jsdom context
Still doesn't work, but I learned a lot. Passed the context all the way
through, but since the VitestExecutor wasn't instantiated with it, it's
also missing this.externalModules. This means VitestExecutor.runModule
will always short circuit and use ViteNodeRunner.runModule.

Also learned about the
experimental.server.config.experimental.skipSsrTransform option in the
Vite server. However, I can't find a way to tweak that from the test
runner.

Of course, this is all seeming more difficult than it needs to be.
Wonder if I should explore the Vitest environment mechanism:

- https://vitest.dev/config/#environment
- https://vitest.dev/guide/environment.html