Add initApp test, fragment helper, page object
The initApp() test uses the jsdom environment provided by Vitest:

- https://vitest.dev/guide/environment.html

The fragment() test helper works as-is in both the jsdom and browser
environments. The StringCalculatorPage object follows the Page Object
pattern from the Selenium WebDriver docs, even though we're not using
Selenium here:

- https://www.selenium.dev/documentation/test_practices/design_strategies/

---

I settled on this after days of trying to figure out how to execute JS
Modules on a JSDOM instance directly. See:

- commit 86ac76f6acc9522f5b271f0ea9096fc5cbb50730
- commit 9c87cbfa82b4bc28306fdcaeaed45885e314def6
- commit d55f78da1f06ea1a4b0f88be58d0be9e950ec864

I finally realized that if what I wanted to do was so difficult, maybe
there was a better way. So this change keeps the PageLoader-based test
from before, to make sure initApp executes on DOMContentLoaded. But for
the new initApp test, as for the upcoming tests for other modules, using
the jsdom environment will suffice.

I was also pleased to discover that `vitest --browser` will do the right
thing and _not_ load the jsdom environment in the browser. So tests
using the jsdom environment will already execute in the browser for
free. (Perhaps there are features which would require a workaround
between the jsdom and browser environments, but I haven't encountered
them yet.)