Pass the <div id="app"> to initApp()
Now main.js retrieves the top level element and passes it to initApp(),
which then gets passed to the other components.

---

While still thinking about the <form> + DocumentFragment problem, it
occurred to me that not every component should need to call
document.querySelector('#app'). I'm also currently expecting the <form>
solution I come up with to require creating elements on the main DOM
instead of a DocumentFragment.

Plus, if we ever want to change that ID, or use a class name, etc., we'd
have to update every querySelector() call. This new architecture
eliminates that particular problem.