Convert static namespace components to classes
Whereas before the pattern was:

  StaticNamespace.init(params)

The pattern is now:

  const c = new ComponentClass()
  c.init(params)

This allows for the possibility of building a proper object graph. See
the implementation comment for App.init() for further explanation.

---

I thought to do this after realizing that modeling everything with
static functions might not suggest the most scalable architecture. While
this example doesn't contain any components that need any state, I
wanted to model more closely what a stateful component graph might look
like.

I thought about doing the same to components/request.js, but I see the
functions in that model as being quite good composable building blocks.
I'd like them to stand alone as an illustrative example.

What I will do in the next commit is create a new Backend abstraction
which uses the request.js functions. It will also allow for requesting a
response from a "Backend" implementation in the App itself. This will
enable this project to serve as a JavaScript environment for the same
String Calculator kata.

The user can then switch between the in-App implementation of the String
Calculator logic and the Tomcat backend. It will also show how the
frontend can run and be independently developed using a local stub
implementation independently from the production backend.