Fix incorrect form.action due to incorrect base
With the base URL '/strcalc', resolving the form action URL './add'
against document.location.href produces the path '/add', which is
incorrect.

With the base URL '/strcalc/', resolving './add' produces
'/strcalc/add' as expected.

---

Writing a test for this directly is difficult. (At least, it is at the
component level using Vitest; an end-to-end Selenium test would catch
it.) Tying the form action URL to a specific absolute path would render
the component and its tests brittle. Also, it seems that trying to
import the config from vite.config.js into a test (to access the
config.base value) breaks the Vitest loader somehow.

As a proxy for a test assertion, the PageLoader constructor now enforces
that the base path ends with '/' (before it enforced that it _didn't_).