Replace Docker with TestTomcat in large tests
With the embedded TestTomcat, there's no real need to use Docker for the
large tests. This should also enable them to run on arm64 Windows,
finally. It also makes the existing large test much faster, since
there's no waiting for Docker to start and stop in the background.

However, I'm leaving the Docker-related testing utils in the repo for
now. I may move them to another repo later, but I think it's handy code
to have for reference.