Get 'pnpm jsdoc' working on Windows
For one thing, Windows can't interpret the `#!/usr/bin/env node` header,
nor can it grok the `bin/jsdoc-cli-wrapper.js` path directly.

Prefixing the script with `node` fixes the second problem. Updating
getPath() to detect the 'win32' platform and adding the .CMD extension
to the candidate path fixes the first. This is because 'pnpm install'
will create a CMD.EXE wrapper for the installed script on Windows.