Improve JS type checking, bump vitest and others
This builds upon my recent experience improving the type checking in:

- mbland/jsdoc-cli-wrapper
- mbland/test-page-opener
- mbland/rollup-plugin-handlebars-precompiler

Switched from jsconfig.json to tsconfig.json, which seems to be better
supported by default. Moved a bunch of `tsc` compiler options to that
file and updated the `pnpm typecheck` script accordingly.

The generated JSDoc output still isn't the greatest, but the Visual
Studio Code and IntelliJ IDEA type checking and documentation is pretty
good now. I may have to circle back to improving the JSDoc output later,
but I need a break from it for now.

Also, bumped vitest to v1.2.1, along with a couple other module
versions.