Call nonexistent ProdStringCalculator.add()
This is the fourth step in the Java Test-Driven Development demo.

Calling this nonexistent method causes the test to fail to build again:

  > Task :strcalc:compileTestJava FAILED
  .../ProdStringCalculatorTest.java:18: error: cannot find symbol
          int result = calc.add("");
                           ^
    symbol:   method add(String)
    location: variable calc of type ProdStringCalculator
  1 error

Again, for the purposes of Test-Driven Development, build failures are
test failures.

Now we're ready to make the test pass again by creating
ProdStringCalculator.add() with just enough logic to pass the assertion.