Add test for adding two numbers
This is the eighth step in the Java Test-Driven Development demo.

The new ProdStringCalculatorTest.returnsSumOfTwoNumbers() fails with an
exception because the current ProdStringCalculator.add() implementation
passes the input string to Integer.parseInt() without breaking it apart:

  > Task :strcalc:test FAILED

  ProdStringCalculatorTest > returnsSumOfTwoNumbers() FAILED
      java.lang.NumberFormatException at ProdStringCalculatorTest.java:36

  4 tests completed, 1 failed

  FAILURE: Build failed with an exception.

Now we're ready to add just enough code to make the test pass again.