Call nonexistent ProdStringCalculator.add()
This is the fourth step in the JavaScript Test-Driven Development demo.

Calling this nonexistent method causes the test to fail to build again:

   ❯ components/prod-calculator.test.js (1)
     ❯ ProdStringCalculator (1)
       × empty string returns zero

  ---- Failed Tests 1 ----

   FAIL  components/prod-calculator.test.js > ProdStringCalculator > empty string returns zero
  TypeError: calc.add is not a function
   ❯ components/prod-calculator.test.js:15:23
       13|     let calc = new ProdStringCalculator()
       14|
       15|     let result = calc.add('')
         |                       ^
       16|
       17|     expect(result).toEqual(0)

  ---- [1/1] ----

   Test Files  1 failed | 6 passed (7)
        Tests  1 failed | 17 passed (18)

Again, for the purposes of Test-Driven Development, build failures are
test failures.

Now we're ready to make the test pass again by creating
ProdStringCalculator.add() with just enough logic to pass the assertion.