Refactor ProdStringCalculator.add
This is the tenth step in the JavaScript Test-Driven Development demo.

The implementation is now a little closer to what it will need to become
for step 2: "Allow the Add method to handle an unknown amount of
numbers." The fact that the tests continue to pass ensure that we
haven't broken existing expectations of the production code's behavior.

Now we're ready to refactor the test itself a bit.